Renamed artworkFetcher to metadata-api
