+++ .idea/gradle.xml
@@ -21,11 +21,11 @@
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
-            <option value="$PROJECT_DIR$/artworkFetcher" />
             <option value="$PROJECT_DIR$/buildSrc" />
             <option value="$PROJECT_DIR$/core" />
             <option value="$PROJECT_DIR$/logger" />
             <option value="$PROJECT_DIR$/media" />
+            <option value="$PROJECT_DIR$/metadata-api" />
             <option value="$PROJECT_DIR$/playback-layers" />
             <option value="$PROJECT_DIR$/playback-layers/database" />
             <option value="$PROJECT_DIR$/testutils" />
