--- artworkFetcher/src/main/java/com/tachyonmusic/artworkfetcher/domain/UrlEncoder.kt
@@ -1,7 +1,0 @@
-package com.tachyonmusic.artworkfetcher.domain
-
-import com.tachyonmusic.util.Resource
-
-interface UrlEncoder {
-    fun encode(baseUrl: String, params: Map<String, String>): Resource<String>
-}
