+++ metadata-api/proguard-rules.pro
