+++ metadata-api/src/main/AndroidManifest.xml
