+++ metadata-api/src/main/java/com/tachyonmusic/metadata_api/domain/UrlEncoder.kt
@@ -1,0 +1,7 @@
+package com.tachyonmusic.metadata_api.domain
+
+import com.tachyonmusic.util.Resource
+
+interface UrlEncoder {
+    fun encode(baseUrl: String, params: Map<String, String>): Resource<String>
+}
