+++ metadata-api/src/main/java/com/tachyonmusic/metadata_api/domain/model/SearchInfo.kt
@@ -1,0 +1,8 @@
+package com.tachyonmusic.metadata_api.domain.model
+
+data class SearchInfo(
+    val title: String? = null,
+    val artist: String? = null,
+    val album: String? = null,
+    val mbid: String? = null
+)
