+++ metadata-api/src/test/java/com/tachyonmusic/metadata_api/data/UrlEncoderTest.kt
@@ -1,0 +1,34 @@
+package com.tachyonmusic.metadata_api.data
+
+import com.tachyonmusic.metadata_api.domain.UrlEncoder
+import com.tachyonmusic.util.Resource
+import org.junit.Test
+
+internal class UrlEncoderTest {
+    val encoders = listOf<UrlEncoder>(
+        UrlEncoderImpl()
+    )
+
+    @Test
+    fun `Empty params, returns Resource Error`() = each {
+        assert(it.encode("BaseUrl.net", mapOf()) is Resource.Error)
+    }
+
+    @Test
+    fun `Valid input, returns correct output`() = each {
+        val urlParams = mapOf(
+            "media" to "music",
+            "entity" to "album",
+            "term" to "Hello World"
+        )
+
+        val encoded = it.encode("BaseUrl.com", urlParams)
+        assert(encoded is Resource.Success)
+        assert(encoded.data == "BaseUrl.com?media=music&entity=album&term=Hello+World")
+    }
+
+    private fun each(action: (UrlEncoder) -> Unit) {
+        for (encoder in encoders)
+            action(encoder)
+    }
+}
