+++ settings.gradle.kts
@@ -20,7 +20,7 @@ include(":media")
 include(":util")
 include(":core")
 include(":testutils")
-include(":artworkFetcher")
+include(":metadata-api")
 include(":logger")
 include(":playback-layers")
 include(":playback-layers:database")
